requirements/base.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-location-field==2.7.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==9.5.0
inflection==0.5.1
marshmallow==3.20.1
packaging==23.1
phonenumberslite==8.13.17
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
